flat bpc graph type
